--- prisma/seed.ts
@@ -1,22 +1,0 @@
-// prisma/seed.ts
-import { PrismaClient } from '@prisma/client';
-
-const prisma = new PrismaClient();
-
-async function main() {
-  // await prisma.user.create({
-  //   data: {
-  //     email: 'trazzle@trazzle.com',
-  //   },
-  // });
-  // 추가적인 데이터 삽입 작업을 여기에 작성할 수 있습니다.
-}
-
-main()
-  .catch((e) => {
-    console.error(e);
-    process.exit(1);
-  })
-  .finally(async () => {
-    await prisma.$disconnect();
-  });
